Refactor tests to split tests with CRuby and tests with JRuby
