+++ src/test/java/org/embulk/util/rubytime/TestRubyDateTimeZonesWithCRuby.java
@@ -1,0 +1,862 @@
+/*
+ * Copyright 2018 The Embulk project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.embulk.util.rubytime;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Stream;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+
+public class TestRubyDateTimeZonesWithCRuby {
+    @ParameterizedTest
+    @MethodSource("provideToOffsetInSecondsParams")
+    public void testToOffsetInSecondsWithCRuby(final String value) throws Exception {
+        assertToOffsetInSecondsCRuby(value);
+    }
+
+    @ParameterizedTest
+    @MethodSource("provideParseOffsetParams")
+    public void testParseOffsetWithCRuby(final String value) throws Exception {
+        assertParseOffsetCRuby(value);
+    }
+
+    private static void assertToOffsetInSecondsCRuby(final String value) throws Exception {
+        // A fraction part of :offset is rounded down.
+        final String oneLiner = String.format(
+                "require 'date'; " +
+                "hashDateTime = DateTime._strptime(\"%s\", \"%%z\"); " +
+                "hashDateTime[:offset] = hashDateTime[:offset].to_i if hashDateTime&.fetch(:offset).is_a?(Rational); " +
+                "puts hashDateTime&.fetch(:offset, nil).inspect; " +
+                "puts hashDateTime&.fetch(:leftover, nil).inspect",
+                value);
+        final List<String> result = cruby.callOneLiner(oneLiner);
+        assertTrue(result.size() == 2);
+
+        final int actual = RubyDateTimeZones.toOffsetInSeconds(value);
+        if (result.get(0).equals("nil") || !result.get(1).equals("nil")) {
+            assertEquals(Integer.MIN_VALUE, actual);
+        } else {
+            assertEquals(Integer.parseInt(result.get(0)), actual);
+        }
+    }
+
+    private static void assertParseOffsetCRuby(final String value) throws Exception {
+        // A fraction part of :offset is rounded down.
+        final String oneLiner = String.format(
+                "require 'date'; " +
+                "hashDateTime = DateTime._strptime(\"%s\", \"%%z\"); " +
+                "hashDateTime[:offset] = hashDateTime[:offset].to_i if hashDateTime&.fetch(:offset).is_a?(Rational); " +
+                "puts hashDateTime&.fetch(:offset, nil).inspect; " +
+                "puts hashDateTime&.fetch(:leftover, nil).inspect",
+                value);
+        final List<String> result = cruby.callOneLiner(oneLiner);
+        assertTrue(result.size() == 2);
+
+        // We don't test parseOffset against strings which will have leftover.
+        // The typical case is a string ending with a colon, such as "UTC+13:".
+        // In such cases, the colon is filtered by the regular expression
+        // before the string is passed to CRuby's date_zone_to_diff.
+        // https://svn.ruby-lang.org/cgi-bin/viewvc.cgi/tags/v2_5_0/ext/date/date_strptime.c?view=markup#l571
+        assertTrue(result.get(1).equals("nil"));
+
+        final int actual = RubyDateTimeZones.parseOffsetForTesting(value);
+        if (result.get(0).equals("nil")) {
+            assertEquals(Integer.MIN_VALUE, actual);
+        } else {
+            assertEquals(Integer.parseInt(result.get(0)), actual);
+        }
+    }
+
+    @BeforeAll
+    public static void loadCRubyCaller() {
+        cruby = new CRubyCaller();
+    }
+
+    private static Stream<String> provideParseOffsetParams() {
+        return Arrays.stream(OFFSETS);
+    }
+
+    private static Stream<String> provideToOffsetInSecondsParams() {
+        return Stream.concat(Arrays.stream(OFFSETS), Arrays.stream(ZONES));
+    }
+
+    private static final String[] OFFSETS = {
+        "DUMMY",
+        "9",
+        "13",
+        "78",
+        "127",
+        "19248",
+        "482844",
+        "9184812",
+        "19184812",
+        "+9",
+        "+13",
+        "+78",
+        "+127",
+        "+19248",
+        "+482844",
+        "+9184812",
+        "+19184812",
+        "-9",
+        "-13",
+        "-78",
+        "-127",
+        "-19248",
+        "-482844",
+        "-9184812",
+        "-19184812",
+        "UTC",
+        "GMT",
+        "UTC+9",
+        "UTC+9:2",
+        "UTC+9:2:3",
+        "UTC+12",
+        "UTC+12:13",
+        "UTC+12:13:14",
+        "UTC+99",
+        "UTC+99:13",
+        "UTC+99:13:14",
+        "UTC+912",
+        "UTC+912:42",
+        "UTC+912:42:49",
+        "UTC+9129",
+        "UTC+91294",
+        "UTC+912942",
+        "UTC+9129424",
+        "UTC+19129424",
+        "UTC+19.5",
+        "UTC+19.25",
+        "UTC+19.125",
+        "UTC+19.0625",
+        "UTC+19.03125",
+        "UTC+19.015625",
+        "UTC+19.0078125",
+        "UTC+19.00390625",
+        "UTC+19.111111",
+        "UTC+19.0111111",
+        "UTC+19.00111111",
+        "UTC+19.000111111",
+        "UTC-9",
+        "UTC-9:2",
+        "UTC-9:2:3",
+        "UTC-12",
+        "UTC-12:13",
+        "UTC-12:13:14",
+        "UTC-99",
+        "UTC-99:13",
+        "UTC-99:13:14",
+        "UTC-912",
+        "UTC-912:42",
+        "UTC-912:42:49",
+        "UTC-9129",
+        "UTC-91294",
+        "UTC-912942",
+        "UTC-9129424",
+        "UTC-19129424",
+        "UTC-19.5",
+        "UTC-19.25",
+        "UTC-19.125",
+        "UTC-19.0625",
+        "UTC-19.03125",
+        "UTC-19.015625",
+        "UTC-19.0078125",
+        "UTC-19.00390625",
+        "UTC-19.111111",
+        "UTC-19.0111111",
+        "UTC-19.00111111",
+        "UTC-19.000111111",
+    };
+
+    // A zone name with a space character is not parsed by itself by Ruby's Date._strptime for the regular expression.
+    // https://svn.ruby-lang.org/cgi-bin/viewvc.cgi/tags/v2_5_0/ext/date/date_strptime.c?view=markup#l571
+    //
+    // Such a zone name is parsed only when suffixed by "standard time" or "daylight time".
+    private static final String[] ZONES = {
+        "dummy",
+        "DUMMY",
+        "DuMmY",
+        "DuMmY DsT",
+        // "dummy zone",
+        // "DUMMY zone",
+        // "DuMMy ZOnE",
+        "dumMY zOne StandarD TimE",
+        "ut",
+        "UT",
+        "Ut",
+        "uT DayliGHt \t TimE",
+        "gmt",
+        "GMT",
+        "gmT",
+        "gMt sTANDaRd\t\tTIMe",
+        "est",
+        "EST",
+        "EsT",
+        "eST\t dST",
+        "edt",
+        "EDT",
+        "edT",
+        "Edt  dst",
+        "cst",
+        "CST",
+        "cst",
+        "csT\tDummy  TIME",
+        "cdt",
+        "CDT",
+        "CDT",
+        "cdt \tsTAnDaRd \tTiME",
+        "mst",
+        "MST",
+        "mSt",
+        "msT\tdsT",
+        "mdt",
+        "MDT",
+        "MdT",
+        "mdT \n\nDst",
+        "pst",
+        "PST",
+        "PST",
+        "PST\tdst",
+        "pdt",
+        "PDT",
+        "PDT",
+        "Pdt dAYLigHT\t  TImE",
+        "a",
+        "A",
+        "a  DST",
+        "b",
+        "B",
+        "b DAyligHT  tImE",
+        "c",
+        "C",
+        "C  StanDarD   tImE",
+        "d",
+        "D",
+        "d Dst",
+        "e",
+        "E",
+        "e\tdst",
+        "f",
+        "F",
+        "f \tdUMmY\t\t\ttiME",
+        "g",
+        "G",
+        "g DUMMy\ttiME",
+        "h",
+        "H",
+        "H\tDumMY  TiMe",
+        "i",
+        "I",
+        "i stAnDarD\t tIME",
+        "k",
+        "K",
+        "k\t duMMY\ntIME",
+        "l",
+        "L",
+        "L  sTAndaRd\tTIMe",
+        "m",
+        "M",
+        "m\n daYLIgHT \ttIME",
+        "n",
+        "N",
+        "N  DuMMy\nTIME",
+        "o",
+        "O",
+        "o  DUMmY TIMe",
+        "p",
+        "P",
+        "p DsT",
+        "q",
+        "Q",
+        "q\ndSt",
+        "r",
+        "R",
+        "r\n\tdUMmy TimE",
+        "s",
+        "S",
+        "s dST",
+        "t",
+        "T",
+        "t\tdAYlIGhT\n tIMe",
+        "u",
+        "U",
+        "u\nDST",
+        "v",
+        "V",
+        "V  \tSTANDarD\n\n tiMe",
+        "w",
+        "W",
+        "W\tDuMmy\ntIme",
+        "x",
+        "X",
+        "X\tDummY\n\nTIME",
+        "y",
+        "Y",
+        "Y DAylIGhT Time",
+        "z",
+        "Z",
+        "Z  DST",
+        "utc",
+        "UTC",
+        "UTC",
+        "UTC  DAylIGht tIME",
+        "wet",
+        "WET",
+        "Wet",
+        "WEt Dst",
+        "at",
+        "AT",
+        "aT",
+        "aT dsT",
+        "brst",
+        "BRST",
+        "brST",
+        "BrST\t sTANDArD   timE",
+        "ndt",
+        "NDT",
+        "ndt",
+        "ndt  sTANdArD TIMe",
+        "art",
+        "ART",
+        "aRt",
+        "art  DUmMY tiMe",
+        "adt",
+        "ADT",
+        "adt",
+        "aDT\ndUMmY\nTimE",
+        "brt",
+        "BRT",
+        "brt",
+        "bRT\t\n\nDaYLighT Time",
+        "clst",
+        "CLST",
+        "ClsT",
+        "clsT\n STANdARD\t\ntime",
+        "nst",
+        "NST",
+        "nsT",
+        "nST  dsT",
+        "ast",
+        "AST",
+        "AST",
+        "AsT sTANDard\n tiMe",
+        "clt",
+        "CLT",
+        "cLT",
+        "ClT  DsT",
+        "akdt",
+        "AKDT",
+        "akDt",
+        "aKdT  StANDARD tIME",
+        "ydt",
+        "YDT",
+        "ydt",
+        "YDT DAyLIghT tiME",
+        "akst",
+        "AKST",
+        "Akst",
+        "aksT \tDuMmY \n TiMe",
+        "hadt",
+        "HADT",
+        "HaDt",
+        "hADT  sTaNdarD \tTiMe",
+        "hdt",
+        "HDT",
+        "HdT",
+        "hDT\n\n\nstaNdArD tImE",
+        "yst",
+        "YST",
+        "YsT",
+        "YST\ndaYliGht TImE",
+        "ahst",
+        "AHST",
+        "ahSt",
+        "aHSt\n DUMmY TiME",
+        "cat",
+        "CAT",
+        "Cat",
+        "CAt\n DAyLiGHt\ntiME",
+        "hast",
+        "HAST",
+        "HASt",
+        "hasT DayLighT timE",
+        "hst",
+        "HST",
+        "HsT",
+        "hST\t dUmMY tImE",
+        "nt",
+        "NT",
+        "Nt",
+        "NT\nSTaNdArd\ntime",
+        "idlw",
+        "IDLW",
+        "idlw",
+        "idlw   Dst",
+        "bst",
+        "BST",
+        "BsT",
+        "BST DUmmY\t tImE",
+        "cet",
+        "CET",
+        "CET",
+        "Cet DAYliGHT\ttIME",
+        "fwt",
+        "FWT",
+        "Fwt",
+        "fWt DaYLIGhT\t\ntiMe",
+        "met",
+        "MET",
+        "MeT",
+        "MET\ndaylIGhT\ttiMe",
+        "mewt",
+        "MEWT",
+        "mEwt",
+        "mEwT\tSTaNdard  tiMe",
+        "mez",
+        "MEZ",
+        "meZ",
+        "MeZ\t\tduMMY\t TimE",
+        "swt",
+        "SWT",
+        "SWt",
+        "swT \nstanDard  \ttIME",
+        "wat",
+        "WAT",
+        "wAT",
+        "WAt   DUMMy \ntiME",
+        "west",
+        "WEST",
+        "WEsT",
+        "WeST\nDuMMy\tTimE",
+        "cest",
+        "CEST",
+        "CESt",
+        "CEsT \nSTanDArd TImE",
+        "eet",
+        "EET",
+        "eet",
+        "eET DuMMY\t\n\ttime",
+        "fst",
+        "FST",
+        "FST",
+        "fSt  DAyLigHT  time",
+        "mest",
+        "MEST",
+        "mESt",
+        "meST\n DumMy \tTiME",
+        "mesz",
+        "MESZ",
+        "Mesz",
+        "MEsz\nduMMY\t\tTiME",
+        "sast",
+        "SAST",
+        "sasT",
+        "Sast \tDuMMy \nTimE",
+        "sst",
+        "SST",
+        "sSt",
+        "sst\n dUmMy time",
+        "bt",
+        "BT",
+        "BT",
+        "Bt\tDaYliGhT\tTImE",
+        "eat",
+        "EAT",
+        "EAt",
+        "Eat\tSTandaRd tIME",
+        "eest",
+        "EEST",
+        "EEst",
+        "EeSt\n\ndUmMy\nTIMe",
+        "msk",
+        "MSK",
+        "MSk",
+        "Msk dAYLiGHT  \ntime",
+        "msd",
+        "MSD",
+        "MSD",
+        "MSd\n\n DAYLigHt\ttiME",
+        // "zp4",
+        // "ZP4",
+        // "ZP4",
+        // "ZP4\nDsT",
+        // "zp5",
+        // "ZP5",
+        // "zp5",
+        // "zP5  stAndaRD tIMe",
+        "ist",
+        "IST",
+        "Ist",
+        "Ist\tDayLigHT\tTIME",
+        // "zp6",
+        // "ZP6",
+        // "Zp6",
+        // "Zp6  DsT",
+        "wast",
+        "WAST",
+        "wASt",
+        "wast \tdSt",
+        "cct",
+        "CCT",
+        "cct",
+        "Cct  dAYlighT Time",
+        "sgt",
+        "SGT",
+        "sgt",
+        "Sgt \tDUmmY\n TImE",
+        "wadt",
+        "WADT",
+        "WaDT",
+        "wADT \ndST",
+        "jst",
+        "JST",
+        "Jst",
+        "jst  DsT",
+        "kst",
+        "KST",
+        "KST",
+        "KSt\n\tDUmMy TiMe",
+        "east",
+        "EAST",
+        "eAst",
+        "eAsT DUMmY\n\ntIme",
+        "gst",
+        "GST",
+        "Gst",
+        "GSt DumMY \n TimE",
+        "eadt",
+        "EADT",
+        "EAdT",
+        "eadT\n\nduMMY  TimE",
+        "idle",
+        "IDLE",
+        "idLe",
+        "IdlE\t dUMmY\n\t TIMe",
+        "nzst",
+        "NZST",
+        "nZst",
+        "NZSt DSt",
+        "nzt",
+        "NZT",
+        "nZT",
+        "nzT Dst",
+        "nzdt",
+        "NZDT",
+        "nZDt",
+        "NzDT\tDST",
+        "afghanistan",
+        "AFGHANISTAN",
+        "afgHAniStaN",
+        "AFGHANIsTAN\n DAyLighT\nTime",
+        "alaskan",
+        "ALASKAN",
+        "alaskan",
+        "AlASKAn DST",
+        "arab",
+        "ARAB",
+        "aRAb",
+        "aRaB  \tSTAndARD  Time",
+        "arabian",
+        "ARABIAN",
+        "aRabIan",
+        "aRaBIAn  DUMMY\ttImE",
+        "arabic",
+        "ARABIC",
+        "ARABIc",
+        "ARabiC\n \tdSt",
+        "atlantic",
+        "ATLANTIC",
+        "AtlANTic",
+        "atlanTic DAyLigHT\ttiMe",
+        // "aus central",
+        // "AUS CENTRAL",
+        // "AUS   CeNTRAl",
+        "AUS CenTraL  DAyLIGht\tTimE",
+        // "aus eastern",
+        // "AUS EASTERN",
+        // "auS  eAsTeRN",
+        "aus\nEastErn\t\nsTaNDarD  Time",
+        "azores",
+        "AZORES",
+        "AzorES",
+        "azores  dUmmy\nTImE",
+        // "canada central",
+        // "CANADA CENTRAL",
+        // "cANADA \nCENTRAl",
+        "CANAda CenTrAl  sTaNdArd\n Time",
+        // "cape verde",
+        // "CAPE VERDE",
+        // "cApe  VERDe",
+        "CAPe  veRde sTandard tIME",
+        "caucasus",
+        "CAUCASUS",
+        "CAuCasUS",
+        "CaucasUs\tdst",
+        // "cen. australia",
+        // "CEN. AUSTRALIA",
+        // "ceN.\t\tausTralIa",
+        "cEN. aUstRALiA StaNdArd  TIMe",
+        // "central america",
+        // "CENTRAL AMERICA",
+        // "centrAL  AmerIca",
+        "CentRal   ameRICA\t\t STandArd tiMe",
+        // "central asia",
+        // "CENTRAL ASIA",
+        // "CeNtRaL  asIa",
+        "CEnTRaL \nasIA dAYLighT  TiME",
+        // "central europe",
+        // "CENTRAL EUROPE",
+        // "CentRaL\n\tEUrOpe",
+        "CEntraL\nEUrOPE STaNDArD TimE",
+        // "central european",
+        // "CENTRAL EUROPEAN",
+        // "CENtrAl \tEurOpeaN",
+        "CeNtral euroPEAN  DAYLiGhT tIMe",
+        // "central pacific",
+        // "CENTRAL PACIFIC",
+        // "CEntral PaCIFic",
+        "CenTRal PACIfic daYlIGHT TiME",
+        "central",
+        "CENTRAL",
+        "Central",
+        "cenTRAL DaYLiGHT TIme",
+        "china",
+        "CHINA",
+        "chiNA",
+        "chIna\t\ndaylIGht\t\ntImE",
+        "dateline",
+        "DATELINE",
+        "dATeLINE",
+        "datElInE DaYLIghT\n  TiME",
+        // "e. africa",
+        // "E. AFRICA",
+        // "E. AfrIcA",
+        "e. AfriCa\nStANDarD time",
+        // "e. australia",
+        // "E. AUSTRALIA",
+        // "e. AUStraLIa",
+        "E.\t\taUSTRALIa\t\tDAylIGHt\ttIME",
+        // "e. europe",
+        // "E. EUROPE",
+        // "E. \t\tEuroPe",
+        "E. \n EUrope\tDummY  TIme",
+        // "e. south america",
+        // "E. SOUTH AMERICA",
+        // "E. SouTH  AMERICA",
+        "e.\t\t SOuTh AmeRIca DAyliGht tImE",
+        "eastern",
+        "EASTERN",
+        "easteRn",
+        "eAsTERN dAyLight TIme",
+        "egypt",
+        "EGYPT",
+        "eGYpT",
+        "EGyPT\t\nstanDARD\nTIME",
+        "ekaterinburg",
+        "EKATERINBURG",
+        "EKATerINBurg",
+        "ekAtERINbURg \tDuMmy\nTIMe",
+        "fiji",
+        "FIJI",
+        "fIJi",
+        "fiji DaYlIght \ttImE",
+        "fle",
+        "FLE",
+        "fLe",
+        "FLe StaNDArD\n Time",
+        "greenland",
+        "GREENLAND",
+        "gREenLAnd",
+        "GReENlAnD \n STAndArd  TimE",
+        "greenwich",
+        "GREENWICH",
+        "gReeNwIcH",
+        "grEeNWIch   Dst",
+        "gtb",
+        "GTB",
+        "gTB",
+        "gtB \tSTAndard \ntiMe",
+        "hawaiian",
+        "HAWAIIAN",
+        "HAWaiIAn",
+        "HawaIiAN\n \tdumMY  tIMe",
+        "india",
+        "INDIA",
+        "India",
+        "inDia\ndaYligHt\tTiME",
+        "iran",
+        "IRAN",
+        "IRan",
+        "Iran\tDummY \ttIME",
+        "jerusalem",
+        "JERUSALEM",
+        "JeRUsaleM",
+        "JerUsALEM\n \nDst",
+        "korea",
+        "KOREA",
+        "KOReA",
+        "kOrEA Dst",
+        "mexico",
+        "MEXICO",
+        "MeXico",
+        "mexICo\tdaYLIGhT\t timE",
+        // "mid-atlantic",
+        // "MID-ATLANTIC",
+        // "mid-aTlANtIC",
+        // "mId-AtLanTIC\tStAnDarD\t TImE",
+        "mountain",
+        "MOUNTAIN",
+        "mOUntaiN",
+        "moUnTaiN \nstaNDarD \ttImE",
+        "myanmar",
+        "MYANMAR",
+        "mYaNmar",
+        "mYaNMAR  stanDard \tTiME",
+        // "n. central asia",
+        // "N. CENTRAL ASIA",
+        // "n.\tCEnTral \n\taSia",
+        "N.  cEnTrAL\nasIA\ndAYliGht TIME",
+        "nepal",
+        "NEPAL",
+        "nepal",
+        "nepal  StAnDARd\tTIME",
+        // "new zealand",
+        // "NEW ZEALAND",
+        // "neW \nZEalAnD",
+        "nEw \t\nZEaLaNd  dummy\t tIME",
+        "newfoundland",
+        "NEWFOUNDLAND",
+        "newFoUnDlAnD",
+        "NEWfounDLAnD daYLIGht  TIME",
+        // "north asia east",
+        // "NORTH ASIA EAST",
+        // "norTH\n aSIa  EaSt",
+        "nOrth \naSiA\n eaST  stAndArd\nTime",
+        // "north asia",
+        // "NORTH ASIA",
+        // "NORth\tAsia",
+        "nORTh\nasIa dUmMy\nTIme",
+        // "pacific sa",
+        // "PACIFIC SA",
+        // "pACifIC Sa",
+        "PACIfIc\tSA\n sTANDarD tIME",
+        "pacific",
+        "PACIFIC",
+        "pAcIFiC",
+        "pAcIFic   dAyLIgHT  tiMe",
+        "romance",
+        "ROMANCE",
+        "rOmancE",
+        "ROMANce DayLIghT\ntiME",
+        "russian",
+        "RUSSIAN",
+        "russiaN",
+        "RUSSian\tDSt",
+        // "sa eastern",
+        // "SA EASTERN",
+        // "sa  easTERN",
+        "Sa\n EAstERN dummY tImE",
+        // "sa pacific",
+        // "SA PACIFIC",
+        // "sa\n paCIFIC",
+        "SA \t PAciFIC  \nSTAndArd\ntiMe",
+        // "sa western",
+        // "SA WESTERN",
+        // "SA westeRN",
+        "Sa  wEstern  DaYLigHT\ntiME",
+        "samoa",
+        "SAMOA",
+        "samOA",
+        "SAMoa\t  DSt",
+        // "se asia",
+        // "SE ASIA",
+        // "se \tASiA",
+        "sE\n\t ASia\tSTANdARD tIME",
+        // "malay peninsula",
+        // "MALAY PENINSULA",
+        // "MALaY PENInSUlA",
+        "MAlay PEnInSuLa\tdayLIgHt\n\ntIme",
+        // "south africa",
+        // "SOUTH AFRICA",
+        // "south afriCa",
+        "SOuTh AFRicA DAYlIGht\t\tTIME",
+        // "sri lanka",
+        // "SRI LANKA",
+        // "SrI lAnKA",
+        "sri lanKa\tDuMmY\n  tImE",
+        "taipei",
+        "TAIPEI",
+        "tAiPei",
+        "taipEi dAyLIght TIMe",
+        "tasmania",
+        "TASMANIA",
+        "tAsMaNIA",
+        "taSMaNIA sTaNDARD time",
+        "tokyo",
+        "TOKYO",
+        "ToKYo",
+        "ToKYO\tDUmmY\ttIme",
+        "tonga",
+        "TONGA",
+        "Tonga",
+        "TONGa  dUMMy \ntimE",
+        // "us eastern",
+        // "US EASTERN",
+        // "Us \nEASterN",
+        "us EaSTerN daYliGHt\n\tTimE",
+        // "us mountain",
+        // "US MOUNTAIN",
+        // "us  MOUNTaIN",
+        "uS MOuNtAin\n DuMMy\n\ttiMe",
+        "vladivostok",
+        "VLADIVOSTOK",
+        "vladIvosTOk",
+        "vLADIVOSToK  dsT",
+        // "w. australia",
+        // "W. AUSTRALIA",
+        // "w.\naUsTrAliA",
+        "w.\n aUstRalIa \tDUMMY  TImE",
+        // "w. central africa",
+        // "W. CENTRAL AFRICA",
+        // "W. \ncENtrAl  AFrIca",
+        "w. CentrAL\tafriCa DAYligHt TIME",
+        // "w. europe",
+        // "W. EUROPE",
+        // "W. EuRopE",
+        "W.  EUrope\n  duMMY  tiMe",
+        // "west asia",
+        // "WEST ASIA",
+        // "WeSt asIa",
+        "wesT  AsIa\n\tSTanDarD TimE",
+        // "west pacific",
+        // "WEST PACIFIC",
+        // "wEST pacIFIc",
+        "wESt\t PAciFIC dUMmY\ntimE",
+        "yakutsk",
+        "YAKUTSK",
+        "YaKuTSk",
+        "YaKutSK  dAyLIgHT  TIMe",
+    };
+
+    private static CRubyCaller cruby;
+}
